pyproject.toml
[project]
name = "uumpa-ai"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.1",
    "kubernetes>=35.0.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[project.scripts]
uai = "uumpa_ai.cli:main"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"